# swagger.yaml
...
swagger: '2.0'
info:
  title: Buildings
  description: Lorem Ipsum is simply dummy text of the printing and typesetting industry.
    Lorem Ipsum has been the industry's standard dummy text ever since the 1500s,
    when an unknown printer took a galley of type and scrambled it to make a type
    specimen book.
  version: 1.0.1
produces:
- application/json
...
host: localhost:3000
basePath: "/api"
definitions:
...
  Reading:
    required:
      - BuildingId
      - ObjectId
      - DataFieldId
      - Value
      - Timestamp
    properties:
      BuildingId:
        type: number
      ObjectId:
        type: number
      DataFieldId:
        type: number
      Value:
        type: string
      Timestamp:
        type: string
paths:
...
  "/reading":
    post:
      tags:
      - Reading
      summary: Add a Reading
      description: Save a new Reading
      parameters:
      - in: body
        name: body
        required: true
        description: Reading object that needs to be added to the store
        schema:
          "$ref": "#/definitions/Reading"
      responses:
        '200':
          description: OK
    put:
      description: Update Reading with id
      tags:
      - Reading
      summary: Update a Reading
      parameters:
      - in: body
        name: body
        required: true
        description: Reading object that needs to be update to the store
        schema:
          "$ref": "#/definitions/Reading"
      responses:
        '200':
          description: OK
    get:
      tags:
      - Reading
      summary: Returns a list of Readings
      description: Optional extended description in CommonMark or HTML.
      responses:
        '200':
          description: A JSON array of Meter names
  "/reading/{id}":
    get:
      tags:
      - Reading
      summary: Returns an Object of Reading
      parameters:
      - in: path
        name: id
        type: string
        required: true
      responses:
        '200':
          description: OK
    delete:
      tags:
      - Reading
      summary: Delete Reading by id
      parameters:
      - in: path
        name: id
        type: string
        required: true
      responses:
        '200':
          description: OK